+++ src/main/java/kitchenpos/dao/MenuProductDao.java
@@ -1,7 +1,7 @@
 package kitchenpos.dao;
 
 import java.util.List;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.domain.menu.MenuProduct;
 
 public interface MenuProductDao {
     MenuProduct save(MenuProduct entity);
