+++ src/main/java/kitchenpos/dao/TableGroupDao.java
@@ -1,7 +1,7 @@
 package kitchenpos.dao;
 
 import java.util.List;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.table.TableGroup;
 
 public interface TableGroupDao {
     TableGroup save(TableGroup entity);
