+++ src/main/java/kitchenpos/dao/repository/TableRepository.java
@@ -3,7 +3,7 @@ package kitchenpos.dao.repository;
 import java.util.List;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.dao.jdbctemplate.JdbcTemplateOrderTableDao;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 import org.springframework.dao.InvalidDataAccessApiUsageException;
 import org.springframework.stereotype.Repository;
