+++ src/main/java/kitchenpos/domain/table/Tables.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.table;
 
 import java.util.List;
 import java.util.Objects;
