+++ src/main/java/kitchenpos/dto/ChangeOrderStatusRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.dto;
 
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.order.OrderStatus;
 
 public class ChangeOrderStatusRequest {
     private OrderStatus orderStatus;
