+++ src/test/java/kitchenpos/domain/MenuProductsTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.math.BigDecimal;
 import java.util.List;
+import kitchenpos.domain.menu.MenuProduct;
+import kitchenpos.domain.menu.MenuProducts;
 import org.junit.jupiter.api.Test;
 
 class MenuProductsTest {
